+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - add line to fully install icon
 - add `jj` bind to Bash
 - add `BLE_DISABLED` variable
+- add Fuchsia cursor theme
 
 ### Changed
 - change installation path of Tera-circle-icon-theme from `/usr/local/share/icons` to `/usr/share/icons`